refactor JDBC store to SnowflakeStore
