+++ ingestion/pom.xml
@@ -265,6 +265,12 @@
       <artifactId>commons-math3</artifactId>
       <version>3.6.1</version>
     </dependency>
+      <dependency>
+          <groupId>dev.feast</groupId>
+          <artifactId>feast-storage-connector-snowflake</artifactId>
+          <version>0.7-SNAPSHOT</version>
+          <scope>compile</scope>
+      </dependency>
 
   </dependencies>
 </project>
