+++ ingestion/src/main/java/feast/ingestion/utils/StoreUtil.java
@@ -24,8 +24,8 @@ import feast.proto.core.StoreProto.Store.StoreType;
 import feast.proto.types.ValueProto.ValueType.Enum;
 import feast.storage.api.writer.FeatureSink;
 import feast.storage.connectors.bigquery.writer.BigQueryFeatureSink;
-import feast.storage.connectors.jdbc.writer.JdbcFeatureSink;
 import feast.storage.connectors.redis.writer.RedisFeatureSink;
+import feast.storage.connectors.snowflake.writer.SnowflakeFeatureSink;
 import java.util.HashMap;
 import java.util.Map;
 import org.slf4j.Logger;
@@ -87,8 +87,8 @@ public class StoreUtil {
         return RedisFeatureSink.fromConfig(store.getRedisConfig());
       case BIGQUERY:
         return BigQueryFeatureSink.fromConfig(store.getBigqueryConfig());
-      case JDBC:
-        return JdbcFeatureSink.fromConfig(store.getJdbcConfig());
+      case Snowflake:
+        return SnowflakeFeatureSink.fromConfig(store.getSnowflakeConfig());
       default:
         throw new RuntimeException(String.format("Store type '%s' is unsupported", storeType));
     }
