+++ protos/feast/core/Store.proto
@@ -99,7 +99,7 @@ message Store {
 
     REDIS_CLUSTER = 4;
 
-    JDBC = 5;
+    Snowflake = 5;
   }
 
   message RedisConfig {
@@ -138,7 +138,7 @@ message Store {
     int32 flush_frequency_seconds = 4;
   }
 
-  message JdbcConfig {
+  message SnowflakeConfig {
     string url = 1;
     string class_name = 2;
     int32 batch_size = 3;
@@ -193,6 +193,6 @@ message Store {
     BigQueryConfig bigquery_config = 12;
     CassandraConfig cassandra_config = 13;
     RedisClusterConfig redis_cluster_config = 14;
-    JdbcConfig jdbc_config = 15;
+    SnowflakeConfig Snowflake_config = 15;
   }
 }
