+++ serving/src/test/resources/application-it.properties
@@ -19,7 +19,7 @@ feast.security.authorization.enabled=false
 
 # For ServingServiceSnowflakeStorageIT
 feast.stores[3].name=historical_snowflake
-feast.stores[3].type=JDBC
+feast.stores[3].type=Snowflake
 feast.stores[3].config.schema=PUBLIC
 feast.stores[3].config.database=DEMO_DB
 feast.stores[3].config.role=ACCOUNTADMIN
