+++ storage/connectors/pom.xml
@@ -17,7 +17,7 @@
     <modules>
         <module>redis</module>
         <module>bigquery</module>
-        <module>jdbc</module>
+        <module>snowflake</module>
     </modules>
 
     <build>
