+++ storage/connectors/snowflake/src/main/java/feast/storage/connectors/snowflake/snowflake/TimestampLimits.java
@@ -1,0 +1,44 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ * Copyright 2018-2020 The Feast Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package feast.storage.connectors.snowflake.snowflake;
+
+import java.sql.Timestamp;
+
+public class TimestampLimits {
+
+  public Timestamp min;
+  public Timestamp max;
+
+  /**
+   * Timestamp limits for point-in-time correct retriever
+   *
+   * @param min
+   * @param max
+   */
+  public TimestampLimits(Timestamp min, Timestamp max) {
+    this.min = min;
+    this.max = max;
+  }
+
+  public Timestamp getMax() {
+    return max;
+  }
+
+  public Timestamp getMin() {
+    return min;
+  }
+}
