+++ storage/connectors/snowflake/src/main/resources/templates/join_featuresets_snowflake.sql
