+++ storage/connectors/snowflake/src/main/resources/templates/single_featureset_pit_join_snowflake.sql
