+++ storage/connectors/snowflake/src/test/resources/mockito-extensions/org.mockito.plugins.MockMaker
